requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.3
matplotlib==3.8.2
numpy==1.26.3
